requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bcrypt==4.0.1
bitarray==2.7.6
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
cssselect==1.2.0
cssutils==2.7.1
cytoolz==0.12.1
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dkimpy==1.1.4
dnspython==2.3.0
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.1.1
frozenlist==1.3.3
hexbytes==0.3.1
idna==3.4
importlib-metadata==6.7.0
jaraco.classes==3.2.3
jeepney==0.8.0
jsonschema==4.17.3
keyring==24.2.0
lru-dict==1.2.0
lxml==4.9.2
more-itertools==9.1.0
multidict==6.0.4
parsimonious==0.9.0
premailer==3.10.0
protobuf==4.23.3
psycopg2==2.9.6
pycparser==2.21
pycryptodome==3.18.0
PyJWT==2.7.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
SecretStorage==3.3.3
sqlparse==0.4.4
toolz==0.12.0
urllib3==2.0.3
web3==6.5.0
websockets==11.0.3
yagmail==0.15.293
yarl==1.9.2
zipp==3.15.0
